Extract data query orchestration
